Adding self.config into translate tests; Moving TranslateDycoreFortranData2Py into its own translate_data.py
